credential-management/fedcm.https.html

<!DOCTYPE html>
<title>Federated Credential Management API basics.</title>
<link rel="help" href="https://wicg.github.io/FedCM">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script type="module">
  import {fedcm_test} from './support/fedcm-helper.js';

  const test_options = {
    federated: {
      providers: [{
        url: 'https://idp.test',
        clientId: '1',
        nonce: '2',
      }]
    }
  };

  fedcm_test(async (t, mock) => {
    mock.returnError("ApprovalDeclined");
    const cred = navigator.credentials.get(test_options);
    return promise_rejects_dom(t, 'AbortError', cred);
  }, "User approval decline should reject the promise.");

</script>
